update(frameArray): update the frameArray structure
Include BME data (temp, humidity, pressure), kalman euler and increase the servo range to int16 instead of uint8.

Make the required changes in main, mcu and NAND driver.

Additionally, in the NAND driver, change the frameArray printing statement to be able to print the float values to a csv file.